cloudstack driver
This is a fully working and tested CloudStack driver for docker
machine. All info needed to use this driver for your (private)
CloudStack setup can be found in the updated README.

Of course there are a lot of use cases (different CloudStack setups)
that could need more and/or other options to make the driver also work
for those use cases. But since this is not predictable and this is
currently tested with standard basic and advanced zones, most use cases
should be covered without additional work.

This PR is a followup of PR #129

Signed-off-by: Sander van Harmelen <sander@xanzy.io>