Fix network timeout preventing machine creation
On an airplane, I attempted to create a machine using the VBox driver,
forgetting that we actually presume that network is available for a call
to Github to get the latest boot2docker release.

Users should not need to have access to the Internet in order to use
local drivers.  This fixes the issue and makes creation of a VM without
network connectivity possible :thumbsup:.

Signed-off-by: Nathan LeClaire <nathan.leclaire@gmail.com>