Add shell function recommendation
I find this a lot easier and more pleasant to use than the other
currently supported methods (`env` and `config`) and may serve us well
as a stopgap until/if we can get true config files for Docker.  It's
definitely worth bringing up to users since the workflow is so nice (I
find that I forget the active machine often and want a way to access it
through the same interface I use to run Docker commands against it).

Signed-off-by: Nathan LeClaire <nathan.leclaire@gmail.com>