Fix removal of machine for Vultr driver
Vultr machines cannot be destroyed within 5 minutes of being created. If Remove() gets called within that timelimit, it would delete the SSH key, but fail to do so for the machine itself. Subsequent Remove() calls would then fail because the SSH key cannot be deleted anymore. The machine needs to be deleted first.

Signed-off-by: Fabio Berchtold <jamesclonk@jamesclonk.ch>