Make swarm env the default on swarm master.
On a machine that is the swarm master, a user is likely to intend to
connect to the Swarm manage process. They might accidentally omit it and
get the docker daemon instead, which can be very confusing if you're new
to swarm. The --engine flag provides the old functionality if needed.

Signed-off-by: Don Petersen <don@donpetersen.net>