Add firewall rule for Ubuntu Provisioner
If the image used has a locked-down OS, docker's port
may not be accessible. This patch simply adds an ssh
command that uses Ubuntu's Uncomplicated Firewall (ufw) to
allow the docker port.

I verified that if the firewall rule already exists, an error
is not returned, and the rc is 0. This may not be the case with
all CLIs, but it is for ufw. This means that just running the
command every time won't be a problem

Also note that this should not interfere the existing
Google Compute Engine firewall capabilities because it is at
the OS-level, whereas the 'Firewalls' API handles traffic
at the network level.

Tested against Ubuntu 14.04 on Digital Ocean and Bluemix
(using the openstack driver).

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>