Convert vmwarefusion bats tests back to `machine`
In #615, the helper function `machine()` was created and used within the
vmwarefusion test suite (amongst others) to wrap a platform-specfic
binary during the test suite.

In #1034, it appears that this name was changed to the normal
docker-machine binary, which can incoke the system-installed one, not
the recently compiled one, potentially testing the incorrect binary.

See:
- https://github.com/docker/machine/pull/615#discussion_r25643553
- https://github.com/docker/machine/pull/1034/files#diff-62c88a2fede921556dbdeead731596b2R14

Signed-off-by: Mike Fiedler <miketheman@gmail.com>