Add Firewall Handling to Provisioners
If, for example, a cloud image has a strict firewall in place,
machine will not complete provisioning because it will not be
able to communicate with the Docker port (2376).

This patch adds a few firewallaction package, which is modeled
after the packageaction package. It also includes the Unbuntu
provisioner logic.

The decision to use different logic for each provisioner was
made after realizing that manipulating iptables would have to
be done slightly differently for each provisioner (since they
store their rules in different files) -- so there would already
have to be some provisioner awareness. Additionally, using the
provisioners' firewall manipulation packages (ufw, firewall-cmd,
etc.) will provide the extra benefit of persisting rules across
reboots more easily.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>