Add --openstack-user-data option to OpenStack driver
This patch adds `--openstack-user-data` option to the `docker-machine create` command.
It allows users to specify a user data for the created OpenStack compute instance,
which is processed by Cloud-Init in the early initialization stage of the instance.

The driver takes the file name in the command line option or from the environment,
reads the file, and then pass its content to gophercloud.

Signed-off-by: Kasumi Fukuda <fukuda.k@pepabo.com>