Fixes #1931
Now this change ensures `config.json` is first written into a temp-file and
rename into the original file, if no errors from the FS. If ENOSPC occurs on write,
when the machine is restarted, the previous `config.json` is still available on startup.

Extended `TestStoreSave()` to check for any residual config files in the
machine directory.

Signed-off-by: Anil Belur <askb23@gmail.com>