Support multiple `--amazonec2-security-group`s.
Maintains support for loading old Driver config.json that use the
singular `SecurityGroupName` field, but also adds a new plural
`SecurityGroupNames` field.

This change also bumps to latest stretchr/testify, in the course of
adding a dep on stretchr/testify/mock to help test the modified
`configureSecurityGroups` code.

The unused `deleteSecurityGroup` is dropped as well.

Signed-off-by: John Sirois <john.sirois@gmail.com>
Signed-off-by: Bill Farner <terasurfer@gmail.com>