exoscale: add support for CoreOS
The egoscale binding for exoscale was limiting the images that could be
selected by a user to Ubuntu one only. Enable the use of arbitrary
images. For example, the following images are now available:

 - centos-6.6
 - centos-7.1
 - coreos-stable-835
 - debian-7
 - debian-8
 - ubuntu-12.04
 - ubuntu-14.04
 - ubuntu-15.10

The default user for non-Ubuntu images are not "ubuntu". Let the user
choose the appropriate user ("core" for CoreOS and "debian" for Debian).

CoreOS user are likely to want the ability to provide custom user data
as it enables the configuration of the etcd cluster. We add an option to
provide custom user data file. Moreover, we remove some of the default
options as they are not supported in CoreOS and are useless on
Debian/Ubuntu:

 - Resizing of rootfs is already done on all images by default.
 - Docker Machine will set the hostname as part of the provisioning.

Signed-off-by: Vincent Bernat <Vincent.Bernat@exoscale.ch>