Generic driver: flags to provide SSH config file
This adds two flags to the `generic` driver:

    --generic-ssh-config-file <path/to/config>

        Use the provided config file with the external SSH client (-F)

    --generic-use-user-ssh-config

        Use ~/.ssh/config as the config file

The flags are mutually exclusive in order to avoid ambiguity.

This also adds the SSHConfigFile parameter to BaseDriver and a
GetSSHConfigFile method to all drivers. It seems like there should
be a way to reduce this duplication, since most of the implementations
are identical (return "/dev/null"). I'm a golang noob, though, so if
you know of a better way to accomplish this, please let me know.

Tried this out in my environment, which requires a `ProxyCommand` to
access a remote site via SSH. Before, the issue described in #3310
occurred, preventing me from using `docker-machine` to configure this
host. With this, the `--generic-use-user-ssh-config` option adds
the config file in place of the previous-default `/dev/null` and all
seems well.

Also added a couple unit test cases.

Fixes #3318.

Signed-off-by: Brett Higgins <brhiggins@arbor.net>