provision/coreos: override existing Docker service
CoreOS is shipping a systemd unit starting Docker through socket
activation. The content of the file is the following (CoreOS 899):

    [Unit]
    Description=Docker Socket for the API
    PartOf=docker.service

    [Socket]
    ListenStream=/var/run/docker.sock
    SocketMode=0660
    SocketUser=docker
    SocketGroup=docker

    [Install]
    WantedBy=sockets.target

The service file provided by CoreOS is the following:

    [Unit]
    Description=Docker Application Container Engine
    Documentation=http://docs.docker.com
    After=docker.socket early-docker.target network.target
    Requires=docker.socket early-docker.target

    [Service]
    EnvironmentFile=-/run/flannel_docker_opts.env
    MountFlags=slave
    LimitNOFILE=1048576
    LimitNPROC=1048576
    ExecStart=/usr/lib/coreos/dockerd daemon --host=fd:// $DOCKER_OPTS $DOCKER_OPT_BIP $DOCKER_OPT_MTU $DOCKER_OPT_IPMASQ

    [Install]
    WantedBy=multi-user.target

Docker Machine is setting up a regular systemd unit service file. The
content is the following:

    [Unit]
    Description=Docker Socket for the API
    After=docker.socket early-docker.target network.target
    Requires=docker.socket early-docker.target

    [Service]
    Environment=TMPDIR=/var/tmp
    EnvironmentFile=-/run/flannel_docker_opts.env
    MountFlags=slave
    LimitNOFILE=1048576
    LimitNPROC=1048576
    ExecStart=/usr/lib/coreos/dockerd daemon --host=unix:///var/run/docker.sock --host=tcp://0.0.0.0:2376 --tlsverify --tlscacert /etc/docker/ca.pem --tlscert /etc/docker/server.pem --tlskey /etc/docker/server-key.pem --label provider=exoscale $DOCKER_OPTS $DOCKER_OPT_BIP $DOCKER_OPT_MTU $DOCKER_OPT_IPMASQ
    Environment=

    [Install]
    WantedBy=multi-user.target

It's mostly a copy/paste of the original service. There are multiple problems:

 1. We have two processes listening to /var/run/docker.sock: systemd
    through socket activation and Docker itself. A few tests show that
    Docker gets all the requests but it could be luck and break at some
    point or under some conditions.

 2. The unit setup by Docker Machine is not enabled on start. After a
    reboot, Docker is not started anymore. If we trigger it locally, it
    starts but blocks because it doesn't use socket activation. Further
    invocations will then work.

 3. The environment is in fact empty because of the empty `Environment=`
    at the end.

This commits integrates Docker a bit better with CoreOS by solving those
problems. We cannot do socket activation with TLS because in
stable channels, CoreOS ships only Docker 1.9. Therefore, we keep socket
activation for the Unix socket but we use a regular socket for the TLS
one. We override the existing service to avoid repeat what's in it. We
add /etc/systemd/system/docker.service.d/machine.conf:

    [Service]
    Environment="DOCKER_OPTS=--host=tcp://0.0.0.0:2376 --tlsverify --tlscacert /etc/docker/ca.pem --tlscert /etc/docker/server.pem --tlskey /etc/docker/server-key.pem --label provider=exoscale"

In the future (Docker 1.10+), we could add
/etc/systemd/system/docker.socket.d/machine.conf:

    [Socket]
    ListenStream=0.0.0.0:2376

And only append `Environment` to the `.service` file.

Note that the way the systemd unit is passed over SSH is quite
fragile. Notably, characters need to be escaped properly and a double
quotes should not be present. This can be problematic for the
environment variables.

Signed-off-by: Vincent Bernat <Vincent.Bernat@exoscale.ch>