provision/systemd: override only ExecStart for docker.service
This is somewhat similar to the latest commit, but for Ubuntu. While
Docker 1.11 as downloaded with Ubuntu is recent enough to switch to
socket activation, this would require to modify both docker.service and
docker.socket and the current implementation of the generic provisioner
only enables one location to be replaced. Therefore, just do as for
CoreOS.

The advantage of this option over copying a whole unit file is to ensure
that important options shipped with Docker are conserved.

Signed-off-by: Vincent Bernat <Vincent.Bernat@exoscale.ch>