provision/coreos: Fix Docker version detection
Parse the output of `docker --version` on the remote host instead of
looking for a program named dockerd.

Making decisions based on the presence or absence of a particular
program is short sighted; there are many reasons why dockerd may or may
not exist in $PATH that have nothing to do with the Docker version.

Signed-off-by: Peter Schultz <peter.schultz@classmarkets.com>