Support building a Linux ARM64/aarch64 binary
In order to support building for Linux on ARM64/aarch64 (64bit) we could easily cross compile Docker Machine. In Go we use "linux/arm64", but on Linux the architecture is called "aarch64" which can be determined with `uname -m`.

Testing a build locally within Docker and only compile an ARM64 binary is easy with this command:

```
USE_CONTAINER=true TARGET_OS=linux TARGET_ARCH=arm64 make build-x
```

It builds the binary `./bin/docker-machine-Linux-aarch64`.

Signed-off-by: Dieter Reuter <dieter.reuter@me.com>