Add mount command for mounting machine directories
This comes in handy, when wanting to use volumes with drivers that
don't support folder sharing (such as KVM, as opposed to VirtualBox)

You will need FUSE and SSHFS installed, in order to use this feature.
The machine will also need a "sftp" binary, but most of them have it.

Signed-off-by: Anders F Björklund <anders.f.bjorklund@gmail.com>