Ensure proper cleanup of dangling resources in case of failure
In AWS EC2 driver we start creating resources that simply get lost in case
of failure during instance creation.

The worst case scenario is when a SpotInstance price is too low to be fulfilled
before the timeout.

In such case we already created a Key Pair, but we are not going to persist
`config.json` due to the failure, and we will never be able to delete such key.

Moreover the SpotInstanceRequest will remain in AWS and if it will be fulfilled
later a new instance will be created outside of docker-machine control.

This patch ensures that a proper cleanup is done in case of failure in
`amazonec2.Create()`

Furthermore `docker-machine rm` is now able to handle properly AWS machines
failed during creation thanks to this two new conditions:

1. Key Pair deletion will not fail in case of missing `KeyPairName`
2. Instance termination will not fail in case of missing `InstanceId`

Signed-off-by: Alessio Caiazza <acaiazza@gitlab.com>