Add monitoring flag for Digital Ocean droplets
Digital Ocean now has the functionality for more extensive monitoring of droplets, which includes the ability to set alerts for droplets. These new features only work if a monitoring script is run on the droplet upon startup, fortunately the Digital Ocean API allows for a flag to be passed through (similar to how `private networking` is switched on) that pre-installs the monitoring script allowing the more in-depth monitoring. I have simply added this flag to `docker-machine`.

Signed-off-by: Steven Cooney <dev@stevenjamescooney.com>