amazonec2: correct mis-handling of 'none' VPC
Without this change,

docker-machine create \
    --driver amazonec2 \
    --amazonec2-security-group parity-security-group \
    --amazonec2-instance-type t2.medium \
    --amazonec2-access-key ... \
    --amazonec2-secret-key ... \
    --amazonec2-root-size 128 \
    parity-node

yields the program output

Error with pre-create check: "unable to find a subnet in the zone: us-e
ast-1a"

which comes from drivers/amazonec2/amazonec2.go:506.

But that message is misleading.  After sprinkling in some
`fmt.Println`s, I discovered that the missing piece of info is not
actually the subnet, but rather the VPC ID, which is "none".

With this change, that same command now yields:

Error setting machine configuration from flags provided: amazonec2 driv
er requires either the --amazonec2-subnet-id or --amazonec2-vpc-id opti
on or an AWS Account with a default vpc-id

Signed-off-by: F. Eugene Aumson <feuGeneA@github.com>