Adds 'openstack-timeout-cloud-init' option to openstack driver
By default, this option is set to 0 (disabled).
In case one wants to wait for cloud-init scripts to finish before
installing Docker, set this option to a positive value in order to
wait for cloud-init scripts to finish (i.e.: test the existance of the
'/var/lib/cloud/instance/boot-finished' file).

For example, this is needed in case your user-data file does a
'package_upgrade: true' (system upgrade). Docker machine needs to wait
for the end of the upgrade before beginning installing docker. If not,
there the apt/yum/... packaging database can be locked by either of the processes
and prevent the other from successfully being executed.

Signed-off-by: Julien Barbot <jubarbot@cisco.com>