Support specifying metadata for google compute instances
Sometimes it is desirable to set a few more metadata flags on a new
instance. While most things can be done with a custom image not all of
them really warrant a full blown copy of an image. Also having to update
image because of some minor additional metadata feels like a wastwe of
time.

Signed-off-by: Andreas Rammhold <andreas@rammhold.de>