Add encrypt EBS option to amazonec2 driver
There is not good way to add encryption to an already
unencryped EBS root volume in AWS. It is often required
to have an ec2 instance with an encrypted EBS volume.
The only reasonable solution is to create the encrypted volume
on instance creation. Therefore to create a machine with EBS
encryption enabled the option must be sent via docker-machine.

Signed-off-by: Lucien Brillant <lucienbrillant@gmail.com>