Improve node vsphere node deletion
Problem: Vsphere node deletion entirely relies on the local directory to
ensure uniqueness during node deletion. Pratically, this means the name
is the only unique identifier used. This can be an issue when the
local dir is being modified out of the bounds of docker machine.

Solution: The vsphere driver now uses the ssh key fingerprints of the
keys to ensure the node being deleted is the intended node.