Disable vhd delete when using managed disks
Disabled logic for deleting vhd and storageaccount create/delete
when using managed disks. Prior, vhd delete logic was executed
for managed disks vm which led to a nil pointer exception. This
would disrupt the deletion process for all resources and lead to
remnant resources in Azure. Now, storage accounts are not created
for managed disks, and all resources are properly deleted whether
managed disks were used or not.