Enable NSGs shared between multiple machines
This commit allows users of the Azure driver to submit an NSG name or ARM resource identifier to be used by multiple machines that are provisioned via Rancher Machine. If a user does not submit an NSG name, machine will default to the legacy option of creating an NSG per machine.